Separate domain and infrastructure
