+++ VacationRental.Api.Tests/GetCalendarTests.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using VacationRental.Api.Models;
+using VacationRental.Domain.Models;
 using Xunit;
 
 namespace VacationRental.Api.Tests
