+++ VacationRental.Api.sln
@@ -9,6 +9,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VacationRental.Api.Tests", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationRental.Unit.Tests", "VacationRental.Unit.Tests\VacationRental.Unit.Tests.csproj", "{2F76ABBD-528F-4D6F-9836-E93B3FC8182C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationRental.Domain", "VacationRental.Domain\VacationRental.Domain.csproj", "{492A4970-890F-4241-BEAF-A18711137625}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationRental.Infrastructure", "VacationRental.Infrastructure\VacationRental.Infrastructure.csproj", "{E50E3D26-860F-48DC-9D64-4E9A1EAD48C5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +31,14 @@ Global
 		{2F76ABBD-528F-4D6F-9836-E93B3FC8182C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2F76ABBD-528F-4D6F-9836-E93B3FC8182C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2F76ABBD-528F-4D6F-9836-E93B3FC8182C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{492A4970-890F-4241-BEAF-A18711137625}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{492A4970-890F-4241-BEAF-A18711137625}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{492A4970-890F-4241-BEAF-A18711137625}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{492A4970-890F-4241-BEAF-A18711137625}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E50E3D26-860F-48DC-9D64-4E9A1EAD48C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E50E3D26-860F-48DC-9D64-4E9A1EAD48C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E50E3D26-860F-48DC-9D64-4E9A1EAD48C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E50E3D26-860F-48DC-9D64-4E9A1EAD48C5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
