+++ VacationRental.Domain/Models/BookingBindingModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace VacationRental.Api.Models
+namespace VacationRental.Domain.Models
 {
     public class BookingBindingModel
     {
