+++ VacationRental.Domain/Models/CalendarBookingViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationRental.Api.Models
+﻿namespace VacationRental.Domain.Models
 {
     public class CalendarBookingViewModel
     {
