+++ VacationRental.Domain/Models/CalendarViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace VacationRental.Api.Models
+namespace VacationRental.Domain.Models
 {
     public class CalendarViewModel
     {
