+++ VacationRental.Domain/Models/ResourceIdViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationRental.Api.Models
+﻿namespace VacationRental.Domain.Models
 {
     public class ResourceIdViewModel
     {
