+++ VacationRental.Infrastructure/VacationRental.Infrastructure.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\VacationRental.Domain\VacationRental.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
